add changes to task 5 LettersInWord
